Move template! into sub-module
